services/image_analyze.py
"""
Image Analysis Service
Handles image processing and OCR functionality
"""

import os
from typing import Optional, Dict, Any
from dotenv import load_dotenv

load_dotenv()

def analyze_image_with_openai(image_path: str) -> Optional[Dict[str, Any]]:
    """
    Analyze image using OpenAI Vision API
    """
    try:
        # Placeholder implementation
        # In a real implementation, you would use OpenAI's vision API
        return {
            "text": "Image analysis not implemented",
            "confidence": 0.0,
            "method": "openai"
        }
    except Exception as e:
        print(f"Error analyzing image with OpenAI: {e}")
        return None

def analyze_image_with_gemini(image_path: str) -> Optional[Dict[str, Any]]:
    """
    Analyze image using Google Gemini Vision API
    """
    try:
        # Placeholder implementation
        # In a real implementation, you would use Gemini's vision API
        return {
            "text": "Image analysis not implemented",
            "confidence": 0.0,
            "method": "gemini"
        }
    except Exception as e:
        print(f"Error analyzing image with Gemini: {e}")
        return None

def extract_text_with_tesseract(image_path: str) -> Optional[str]:
    """
    Extract text from image using Tesseract OCR
    """
    try:
        # Placeholder implementation
        # In a real implementation, you would use pytesseract
        return "OCR not implemented"
    except Exception as e:
        print(f"Error extracting text with Tesseract: {e}")
        return None

