requirements.txt
# DocuMind AI - Minimal Requirements for Render Deployment
# This file removes Pinecone temporarily to avoid package conflicts

# Core FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
python-dotenv>=1.0.0

# Essential AI libraries
openai>=1.40.0
google-generativeai>=0.3.0

# Text processing
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0

# HTTP client
httpx>=0.24.0
requests>=2.30.0

# Data processing
numpy>=1.21.0
pandas>=2.0.0

# Document processing
PyMuPDF>=1.23.0
python-docx>=1.1.0
Pillow>=10.0.0
openpyxl>=3.1.0
python-pptx>=0.6.20
mail-parser>=3.15.0
pytesseract>=0.3.10

# Search
rank-bm25>=0.2.2

# Utilities
tldextract>=5.1.0